requirements.txt
asgiref==3.8.1
Django==5.1.4
django-multiselectfield==0.1.13
sqlparse==0.5.3
tzdata==2024.2
